test: 컴포넌트 스토리북 테스트 코드 수정
