+++ src/components/CheckBox/CheckBox.stories.tsx
@@ -1,9 +1,20 @@
+import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
 import CheckBox from './CheckBox';
 
 const meta = {
   title: 'CheckBox',
   component: CheckBox,
+  decorators: [
+    (Story) => {
+      const [isChecked, setIsChecked] = React.useState(false);
+      const onClick = () => {
+        setIsChecked(!isChecked);
+      };
+
+      return <div>{<Story args={{ isCheck: isChecked, onClick }} />}</div>;
+    },
+  ],
 } satisfies Meta<typeof CheckBox>;
 
 export default meta;
@@ -11,5 +22,8 @@ export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {
-  args: {},
+  args: {
+    isCheck: false,
+    onClick: () => {},
+  },
 };
