+++ src/components/ProductList/ProductList.stories.tsx
@@ -1,9 +1,19 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import ProductList from './ProductList';
+import { RecoilRoot } from 'recoil';
 
 const meta = {
   title: 'ProductList',
   component: ProductList,
+  decorators: [
+    (Story) => {
+      return (
+        <RecoilRoot>
+          <Story />
+        </RecoilRoot>
+      );
+    },
+  ],
 } satisfies Meta<typeof ProductList>;
 
 export default meta;
